requirements.txt
python-binance
TA-Lib
numby
websocket_client